% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diversity.R
\name{diversity}
\alias{diversity}
\title{Diversity Stats}
\usage{
diversity(newmap, newtable)
}
\description{
This function runs stats for alpha and beta diversity
}
\examples{
diversity(outtab$newmap,outtab$newtable)
}
