% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Make_Taxa_Tables.R
\docType{data}
\name{domain}
\alias{domain}
\title{Make Taxa Tables}
\format{
An object of class \code{NULL} of length 0.
}
\usage{
domain
}
\arguments{
\item{newmap}{(Required).
table of metadata processed by Make_Tables.}
}
\description{
This function creates taxa tables from dada2 output
}
\keyword{datasets}
