% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Diversity_Plots.R
\docType{data}
\name{kelly_colors}
\alias{kelly_colors}
\title{Diversity Plots}
\format{
An object of class \code{character} of length 21.
}
\usage{
kelly_colors
}
\description{
This function creates plots that summarize alpha and beta diversity
}
\examples{
Diversity_Plots(brayWmeta, outtab$newmap)
}
\keyword{datasets}
