% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Make_Tables.R
\name{Make_Tables}
\alias{Make_Tables}
\title{Make Tables}
\usage{
Make_Tables(inputtable, metadata, taxa)
}
\description{
This function processes dada2 output and returns tables for plots and stats
}
\examples{
# input dada2 sequence table
t1 <- readRDS('seqtab_nochim.rds')
# input metadata
t2 <- read.table('metadata.txt', sep = '\t', comment='', head=TRUE, row.names=1, check.names = FALSE)
# input taxonomy
t3 <- readRDS("taxID.rds")
outtab <- Make_Tables(t1,t2,t3)
}
