% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Make_Taxa_Tables.R
\name{Make_Taxa_Tables}
\alias{Make_Taxa_Tables}
\title{Make Taxa Tables}
\usage{
Make_Taxa_Tables(x)
}
\arguments{
\item{newmap}{(Required).
table of metadata processed by Make_Tables.}
}
\description{
This function creates taxa tables from dada2 output
}
