% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sequence_count_plot.R
\name{sequence_count_plot}
\alias{sequence_count_plot}
\title{sequence count plot}
\usage{
sequence_count_plot(sct)
}
\arguments{
\item{sct}{(Required).
the table containing data to plot.}
}
\description{
this function creates a summary plot of sequence counts
}
\examples{
sequence_count_table <- read.delim("sequence_process_summary.txt", row.names=1)
sequence_count_plot(sequence_count_table)
}
