% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Taxonomy_Plots.R
\name{Taxonomy_Plots}
\alias{Taxonomy_Plots}
\title{Taxonomy Plots}
\usage{
Taxonomy_Plots(meta)
}
\description{
This function creates plots that summarize prevalent taxa
}
\examples{
Taxonomy_Plots(outtab$newtable)
}
