% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diversity.R
\name{diversity}
\alias{diversity}
\title{Diversity Stats}
\usage{
diversity(newmap, newtable)
}
\arguments{
\item{newmap}{(Required).
table of metadata.}

\item{newtable}{(Required).
table of data.}
}
\description{
This function runs stats for alpha and beta diversity
}
