% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Taxonomy_Plots.R, R/sequence_count_plot.R
\docType{data}
\name{value}
\alias{value}
\title{Taxonomy Plots}
\format{
An object of class \code{NULL} of length 0.

An object of class \code{NULL} of length 0.
}
\usage{
value

value
}
\arguments{
\item{meta}{(Required)
table of metadata.}

\item{sct}{(Required).
the table containing data to plot.}
}
\description{
This function creates plots that summarize prevalent taxa

this function creates a summary plot of sequence counts
}
\keyword{datasets}
